+++ src/main/java/com/ss6051/backendspring/global/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package com.ss6051.backendspring.auth.exception;
+package com.ss6051.backendspring.global.exception;
 
 public class UnauthorizedException extends RuntimeException {
     public UnauthorizedException(String message) {
